Feature: Update to Bitmap background and remove the 24 text layers
A screenshot of the application was used and modified so that there
was no time representation on there.

The background image is loaded into a BMP layer.

Note to future self:

    resource_init_current_app(&APP_RESOURCES);

This must be called before a resource is used.  The APP_RESOURCES constant
refers to  the versionDefName variable in the resources json.